Replace test with newtest.  Misc cleanups.
