+++ main.c
@@ -1,5 +1,5 @@
 /*
- * main.c
+ * newtest.c
  *
  * Copyright (c) 2014 Jeremy Garff <jer @ jers.net>
  *
@@ -28,6 +28,8 @@
  */
 
 
+static char VERSION[] = "testing...";
+
 #include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -38,6 +40,9 @@
 #include <fcntl.h>
 #include <sys/mman.h>
 #include <signal.h>
+#include <stdarg.h>
+#include <getopt.h>
+
 
 #include "clk.h"
 #include "gpio.h"
@@ -49,16 +54,23 @@
 
 #define ARRAY_SIZE(stuff)                        (sizeof(stuff) / sizeof(stuff[0]))
 
+// defaults for cmdline options
 #define TARGET_FREQ                              WS2811_TARGET_FREQ
 #define GPIO_PIN                                 18
 #define DMA                                      5
-//#define STRIP_TYPE				WS2811_STRIP_GBR		// WS2812/SK6812RGB integrated chip+leds
-//#define STRIP_TYPE				SK6812_STRIP_RGBW		// SK6812RGBW (NOT SK6812RGB)
+#define STRIP_TYPE				                 WS2811_STRIP_RGB		// WS2812/SK6812RGB integrated chip+leds
+//#define STRIP_TYPE				             WS2811_STRIP_GBR		// WS2812/SK6812RGB integrated chip+leds
+//#define STRIP_TYPE				             SK6812_STRIP_RGBW		// SK6812RGBW (NOT SK6812RGB)
 
 #define WIDTH                                    8
 #define HEIGHT                                   8
 #define LED_COUNT                                (WIDTH * HEIGHT)
 
+int width = WIDTH;
+int height = HEIGHT;
+int led_count = LED_COUNT;
+
+int clear_on_exit = 0;
 
 ws2811_t ledstring =
 {
@@ -72,9 +84,7 @@ ws2811_t ledstring =
             .count = LED_COUNT,
             .invert = 0,
             .brightness = 255,
-#ifdef STRIP_TYPE
             .strip_type = STRIP_TYPE,
-#endif
         },
         [1] =
         {
@@ -86,7 +96,7 @@ ws2811_t ledstring =
     },
 };
 
-ws2811_led_t matrix[WIDTH][HEIGHT];
+ws2811_led_t *matrix;
 
 static uint8_t running = 1;
 
@@ -94,11 +104,11 @@ void matrix_render(void)
 {
     int x, y;
 
-    for (x = 0; x < WIDTH; x++)
+    for (x = 0; x < width; x++)
     {
-        for (y = 0; y < HEIGHT; y++)
+        for (y = 0; y < height; y++)
         {
-            ledstring.channel[0].leds[(y * WIDTH) + x] = matrix[x][y];
+            ledstring.channel[0].leds[(y * width) + x] = matrix[y * width + x];
         }
     }
 }
@@ -107,11 +117,24 @@ void matrix_raise(void)
 {
     int x, y;
 
-    for (y = 0; y < (HEIGHT - 1); y++)
+    for (y = 0; y < (height - 1); y++)
     {
-        for (x = 0; x < WIDTH; x++)
+        for (x = 0; x < width; x++)
         {
-            matrix[x][y] = matrix[x][y + 1];
+            matrix[y * width + x] = matrix[(y + 1)*width + x];
+        }
+    }
+}
+
+void matrix_clear(void)
+{
+    int x, y;
+
+    for (y = 0; y < (height ); y++)
+    {
+        for (x = 0; x < width; x++)
+        {
+            matrix[y * width + x] = 0;
         }
     }
 }
@@ -119,14 +142,27 @@ void matrix_raise(void)
 int dotspos[] = { 0, 1, 2, 3, 4, 5, 6, 7 };
 ws2811_led_t dotcolors[] =
 {
-    0x200000,  // red
-    0x201000,  // orange
-    0x202000,  // yellow
-    0x002000,  // green
-    0x002020,  // lightblue
-    0x000020,  // blue
-    0x100010,  // purple
-    0x200010,  // pink
+    0x00200000,  // red
+    0x00201000,  // orange
+    0x00202000,  // yellow
+    0x00002000,  // green
+    0x00002020,  // lightblue
+    0x00000020,  // blue
+    0x00100010,  // purple
+    0x00200010,  // pink
+};
+
+ws2811_led_t dotcolors_rgbw[] =
+{
+    0x00200000,  // red
+    0x10200000,  // red + W
+    0x00002000,  // green
+    0x10002000,  // green + W
+    0x00000020,  // blue
+    0x10000020,  // blue + W
+    0x00101010,  // white
+    0x10101010,  // white + W
+
 };
 
 void matrix_bottom(void)
@@ -136,12 +172,16 @@ void matrix_bottom(void)
     for (i = 0; i < ARRAY_SIZE(dotspos); i++)
     {
         dotspos[i]++;
-        if (dotspos[i] > (WIDTH - 1))
+        if (dotspos[i] > (width - 1))
         {
             dotspos[i] = 0;
         }
 
-        matrix[dotspos[i]][HEIGHT - 1] = dotcolors[i];
+	if (ledstring.channel[0].strip_type == SK6812_STRIP_RGBW) {
+		matrix[dotspos[i] + (height - 1) * width] = dotcolors_rgbw[i];
+	} else {
+		matrix[dotspos[i] + (height - 1) * width] = dotcolors[i];
+	}
     }
 }
 
@@ -161,9 +201,183 @@ static void setup_handlers(void)
     sigaction(SIGTERM, &sa, NULL);
 }
 
+
+void parseargs(int argc, char **argv, ws2811_t *ws2811)
+{
+	int index;
+	int c;
+
+	static struct option longopts[] =
+	{
+		{"help", no_argument, 0, 'h'},
+		{"dma", required_argument, 0, 'd'},
+		{"gpio", required_argument, 0, 'g'},
+		{"invert", no_argument, 0, 'i'},
+		{"clear", no_argument, 0, 'i'},
+		{"strip", required_argument, 0, 's'},
+		{"height", required_argument, 0, 'y'},
+		{"width", required_argument, 0, 'x'},
+		{"version", no_argument, 0, 'v'},
+		{0, 0, 0, 0}
+	};
+
+	while (1)
+	{
+
+		index = 0;
+		c = getopt_long(argc, argv, "cd:g:his:vx:y:", longopts, &index);
+
+		if (c == -1)
+			break;
+
+		switch (c)
+		{
+		case 0:
+			/* handle flag options (array's 3rd field non-0) */
+			break;
+
+		case 'h':
+			fprintf(stderr, "%s version %s\n", argv[0], VERSION);
+			fprintf(stderr, "Usage: %s \n"
+				"-h (--help)    - this information\n"
+				"-s (--strip)   - strip type - rgb, grb, gbr, rgbw\n"
+				"-x (--width)   - matrix width (default 8)\n"
+				"-y (--height)  - matrix height (default 8)\n"
+				"-d (--dma)     - dma channel to use (default 5)\n"
+				"-g (--gpio)    - GPIO to use must be one of 10,18,40,52\n"
+				"                 If omitted, default is 18\n"
+				"-i (--invert)  - invert pin output (pulse LOW)\n"
+				"-c (--clear)   - clear matrix on exit.\n"
+				"-v (--version) - version information\n"
+				, argv[0]);
+			exit(-1);
+
+		case 'D':
+			break;
+
+		case 'g':
+			if (optarg) {
+				int gpio = atoi(optarg);
+/*
+	https://www.raspberrypi.org/forums/viewtopic.php?f=91&t=105044
+	PWM0, which can be set to use GPIOs 12, 18, 40, and 52. 
+	Only 12 (pin 32) and 18 (pin 12) are available on the B+/2B
+	PWM1 which can be set to use GPIOs 13, 19, 41, 45 and 53. 
+	Only 13 is available on the B+/2B, on pin 35
+*/
+				switch (gpio) {
+					case 10:
+					case 18:
+					case 40:
+					case 52:
+						ws2811->channel[0].gpionum = gpio;
+						break;
+					default:
+						printf ("gpio %d doesnt support PWM0\n",gpio);
+						exit (-1);
+				}
+			}
+			break;
+
+		case 'i':
+			ws2811->channel[0].invert=1;
+			break;
+
+		case 'c':
+			clear_on_exit=1;
+			break;
+
+		case 'd':
+			if (optarg) {
+				int dma = atoi(optarg);
+				if (dma < 14) {
+					ws2811->dmanum = dma;
+				} else {
+					printf ("invalid dma %d\n", dma);
+					exit (-1);
+				}
+			}
+			break;
+
+		case 'y':
+			if (optarg) {
+				height = atoi(optarg);
+				if (height > 0) {
+					ws2811->channel[0].count = height * width;
+				} else {
+					printf ("invalid height %d\n", height);
+					exit (-1);
+				}
+			}
+			break;
+
+		case 'x':
+			if (optarg) {
+				width = atoi(optarg);
+				if (width > 0) {
+					ws2811->channel[0].count = height * width;
+				} else {
+					printf ("invalid width %d\n", width);
+					exit (-1);
+				}
+			}
+			break;
+
+		case 's':
+			if (optarg) {
+				if (!strncasecmp("rgb", optarg, 4)) {
+					ws2811->channel[0].strip_type = WS2811_STRIP_RGB;
+				}
+				else if (!strncasecmp("rbg", optarg, 4)) {
+					ws2811->channel[0].strip_type = WS2811_STRIP_RBG;
+				}
+				else if (!strncasecmp("grb", optarg, 4)) {
+					ws2811->channel[0].strip_type = WS2811_STRIP_GRB;
+				}
+				else if (!strncasecmp("gbr", optarg, 4)) {
+					ws2811->channel[0].strip_type = WS2811_STRIP_GBR;
+				}
+				else if (!strncasecmp("brg", optarg, 4)) {
+					ws2811->channel[0].strip_type = WS2811_STRIP_BRG;
+				}
+				else if (!strncasecmp("bgr", optarg, 4)) {
+					ws2811->channel[0].strip_type = WS2811_STRIP_BGR;
+				}
+				else if (!strncasecmp("rgbw", optarg, 4)) {
+					ws2811->channel[0].strip_type = SK6812_STRIP_RGBW;
+				}
+				else if (!strncasecmp("grbw", optarg, 4)) {
+					ws2811->channel[0].strip_type = SK6812_STRIP_GRBW;
+				}
+				else {
+					printf ("invalid strip %s\n", optarg);
+					exit (-1);
+				}
+			}
+			break;
+
+		case 'v':
+			fprintf(stderr, "%s version %s\n", argv[0], VERSION);
+			exit(-1);
+
+		case '?':
+			/* getopt_long already reported error? */
+			exit(-1);
+
+		default:
+			exit(-1);
+		}
+	}
+}
+
+
 int main(int argc, char *argv[])
 {
     int ret = 0;
+
+    parseargs(argc, argv, &ledstring);
+
+    matrix = malloc(sizeof(ws2811_led_t) * width * height);
 
     setup_handlers();
 
@@ -188,8 +402,15 @@ int main(int argc, char *argv[])
         usleep(1000000 / 15);
     }
 
+    if (clear_on_exit) {
+	matrix_clear();
+	matrix_render();
+	ws2811_render(&ledstring);
+    }
+
     ws2811_fini(&ledstring);
 
+    printf ("\n");
     return ret;
 }
